Enforce dynamic borrow checking of this_cpu
There are many references to this_cpu() and this_cpu_mut() across the
code base but the compiler is unable to perform borrow checking of these
references because it doesn't recognize that they access the same
structure.  This change extends this_cpu() and this_cpu_mut() to perform
dynamic borrow checking to enforce the standard Rust borrow conflict rules.
This borrow checking is performed only in debug builds.

[jroedel: Fixed release builds]

Co-developed-by: Joerg Roedel <jroedel@suse.com>
Signed-off-by: Jon Lange <jlange@microsoft.com>