Access GHCB separately from this_cpu
The GHCB is a separate data structure which can be managed separately
from borrow checking of the current CPU.  However, because the GHCB is
used by reentrant code, it is necessary to provide nesting behavior
which does not conform to the standard borrow checking rules.  This
change separates out references to the GHCB so they do not require
references to the current CPU.  Borrow checking and nesting logic will
come in a future change.

[jroedel: Fixed test-in-svsm]

Co-developed-by: Joerg Roedel <jroedel@suse.com>
Signed-off-by: Jon Lange <jlange@microsoft.com>