kernel: implement boot kernel page heap
This change permits use of the kernel page heap for allocations during
stage2 to hold data that will be consumed by the kernel.  This removes
the awkward logic that previously existed to create a kernel copy of the
IGVM parameters that was embedded in the kernel area between the kernel
ELF image and the page heap, and instead permits the use of a generic
allocator to accomplish the same allocation.

Signed-off-by: Jon Lange <jlange@microsoft.com>